+++ pulse-binding/src/context/subscribe.rs
@@ -66,10 +66,10 @@ use crate::callbacks::{box_closure_get_capi_ptr, MultiUseCallback};
 pub use capi::context::subscribe::pa_subscription_event_type_t as EventType;
 
 /// Mask to extract facility value from the event type passed to the user callback.
-#[deprecated(note="use the associated constant on `Facility` instead")]
+#[deprecated(since = "2.20.0", note="use the associated constant on `Facility` instead")]
 pub const FACILITY_MASK: u32 = capi::PA_SUBSCRIPTION_EVENT_FACILITY_MASK;
 /// Mask to extract operation value from the event type passed to the user callback.
-#[deprecated(note="use the associated constant on `Operation` instead")]
+#[deprecated(since = "2.20.0", note="use the associated constant on `Operation` instead")]
 pub const OPERATION_MASK: u32 = capi::PA_SUBSCRIPTION_EVENT_TYPE_MASK;
 
 bitflags! {
@@ -108,7 +108,7 @@ bitflags! {
 }
 
 /// A set of masks used for expressing which facilities you are interested in when subscribing.
-#[deprecated(note = "Use the associated constants on `InterestMaskSet`.")]
+#[deprecated(since = "2.20.0", note = "Use the associated constants on `InterestMaskSet`.")]
 #[allow(missing_docs)]
 pub mod subscription_masks {
     use super::InterestMaskSet;