+++ pulse-binding/src/direction.rs
@@ -29,7 +29,7 @@ bitflags! {
 }
 
 /// Available flags for `FlagSet`.
-#[deprecated(note = "Use the associated constants on `FlagSet`.")]
+#[deprecated(since = "2.20.0", note = "Use the associated constants on `FlagSet`.")]
 pub mod flags {
     use super::FlagSet;