+++ pulse-binding/src/mainloop/events/io.rs
@@ -39,7 +39,7 @@ bitflags! {
 }
 
 /// Flags for `FlagSet`.
-#[deprecated(note = "Use the associated constants on `FlagSet`.")]
+#[deprecated(since = "2.20.0", note = "Use the associated constants on `FlagSet`.")]
 pub mod flags {
     use super::FlagSet;